Add policies for likes

+++ app/controllers/likes_controller.rb
@@ -1,6 +1,7 @@
 class LikesController < ApplicationController
   def create
     @product = Product.find(params[:product_id])
+    authorize @product.likes
     @product.likes.create(user_id: current_user.id)
 
     redirect_back fallback_location: products_path
@@ -8,6 +9,7 @@ class LikesController < ApplicationController
 
   def destroy
     like = Like.find(params[:id])
+    authorize like
     like.destroy
 
     redirect_back fallback_location: products_path
